Split tests into separated files (#23)
