+++ tests/mappings/modules/core/vat/handleSuck.test.ts
@@ -1,9 +1,9 @@
 import { Bytes, BigInt, BigDecimal } from "@graphprotocol/graph-ts";
 import { test, assert, clearStore } from "matchstick-as";
-import { User } from "../../../../generated/schema";
-import { LogNote } from "../../../../generated/Vat/Vat";
-import { handleSuck } from "../../../../src/mappings/modules/core/vat";
-import { tests } from "../../../../src/mappings/modules/tests";
+import { User } from "../../../../../generated/schema";
+import { LogNote } from "../../../../../generated/Vat/Vat";
+import { handleSuck } from "../../../../../src/mappings/modules/core/vat";
+import { tests } from "../../../../../src/mappings/modules/tests";
 
 function createEvent(u: string, v: string, rad: string): LogNote {
   let sig = tests.helpers.params.getBytes("sig", Bytes.fromHexString("0x1a0b287e"));
